# Manually inline setBitmapRange to *bitmapContainer.lazyIOR function