Even though it adds code duplication the performance benchmarks make
this approach interesting:

```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkRealDataFastOr/census-income_srt-8          617026        395117        -35.96%
BenchmarkRealDataFastOr/census1881_srt-8             1051061       935098        -11.03%
BenchmarkRealDataFastOr/weather_sept_85_srt-8        404418        297734        -26.38%
BenchmarkRealDataFastOr/wikileaks-noquotes_srt-8     214869        190827        -11.19%
```

Note that the benchmarks were done against FastOr in #143