@lemire 2018-02-06
I think that Go allows for manual inlining... so this looks good to me.