@maciej 2018-02-09
Although the Go compiler does function inlining it does not inline functions that are "complex" (that for example contain loops[1]) like `setBitmapRange`. Since in practice the time needed to invoke `setBitmapRange` function was roughly similar to the time `setBitmapRange` executed it made sense to copy the code into the `lazyIOR` function to improve performance (at a cost of code duplication).

[1] https://github.com/golang/go/wiki/CompilerOptimizations#function-inlining